feat: add rebornplusplus/chisel-tools
The chisel-tools go package contains command line tool(s) that helps
working with chisel and chisel slices. In it's current state, the "sdf"
utility from the chisel-tools helps verify the installation of slices
from slice definition files. This can replace some of the CI in the
canonical/chisel-releases repository.

The go package is copied from github.com/rebornplusplus/chisel-tools.
Furthermore, the name is changed to
github.com/canonical/rocks-toolbox/chisel-tools. Once this commit is
merged, we will need to register this package in pkg.go.dev.